Add setup method to detect proxy settings. Extract setup methods into AbstractTest class.
